feat(nomina): cimientos de la feature 005 - entidades, esquema, semillas y permisos
Fases 1 y 2 de specs/005-nomina-novedades-liquidacion (T001-T035, T054):
enums y codigos de parametros legales; 14 entidades nuevas (planes, conceptos
con vigencia, parametros legales con tablas, novedades, corridas, pagos,
entregas) con sus configuraciones EF; PayPeriod y Employee extendidos;
migracion en par NominaNovedadesYLiquidacion con relleno idempotente del plan
DEFAULT; semillas de plan, ~40 conceptos, parametros legales 2026 y tipo de
comprobante NM; 22 permisos Payroll.* y su reparto en los roles integrados;
EmailMessage con adjuntos; retirado el calculo preliminar heredado.

Verificado: AutoMigrate local sobre la base operativa y dos cooperativas;
Architecture.Tests 54/54.

+++ src/Core/IngenIA365ERP.Application/Payroll/PayrollProcessing/Queries/PayrollQueries.cs
@@ -83,7 +83,7 @@ public class ListPayrollSummaryQueryHandler(IApplicationDbContext context)
             deductions,
             net,
             employeeCount,
-            period.Status,
+            (int)period.Status,
             period.StatusMessage));
     }
 }
